2009-01-12  Al Riddoch  <alriddoch@googlemail.com>
	* physics/Vector3D.h: Use sqrtf to ensure efficient use of floats.

+++ physics/Vector3D.h
@@ -52,7 +52,7 @@ float squareDistance(const Point3D & u, const Point3D & v);
 /// Find the distance between two points
 inline float distance(const Point3D & u, const Point3D & v)
 {
-    return sqrt(squareDistance(u, v));
+    return sqrtf(squareDistance(u, v));
 }
 
 typedef std::vector<Point3D> CoordList;
